+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Enh #15: Add trait `HasTemplateItems` class (@terabytesoftw)
 - Enh #16: Add trait `HasTemplateLinkItem` class (@terabytesoftw)
 - Bug #17: Better naming `HasTemplateItems` to `HasTemplateItem` (@terabytesoftw)
+- Bug #18: Add `linkTag()` method in `HasLinkCollection` trait (@terabytesoftw)
 
 ## 0.1.0 March 5, 2024